+++ pizza-byte/pizza-byte.contracts/pizza-byte/PutPizzaRequest.cs
@@ -5,5 +5,5 @@ public record PutPizzaRequest(
     List<string> Toppings,
     string Crust,
     string Size,
-    double Price
+    decimal Price
 );
